[REFACOTOR] event-listener.js로 분리 완료
